Fix nll_backward for negative weights.
Fixes https://github.com/pytorch/pytorch/issues/64256
It adds a test with negative weights as well.

This PR also cleans the code of `nll_backward` which contained a number
of inconsistencies between the different branches of the code: some of
them would check that an index was correct while others didn't. Now the
code is more homogeneous and hopefully easier to follow.

ghstack-source-id: 87d212b012d374f1d023a409be6098e7f648d11c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/64275